+++ src/platform/linux/linux_dgeom.h
@@ -30,6 +30,6 @@ size_t __iomin(char *path, int fd);
  * @param path
  * @return
  */
-size_t __linux_dgeom(struct devgeom *dgeom, char *path);
+int __linux_dgeom(struct devgeom *dgeom, char *path);
 
 #endif //_LINUX_DGEOM_H_
